- Added new component for add todo - Added list component - Renemed folders by camelcase
